added some yaml configurations for the authentication
